# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación API WaaS",
        "description": "Documentación de la API de WaaS. Es necesario contar con un token de autenticación para poder acceder a los recursos de la API. \r\n    *       Para ello se debe enviar el token en el header de la petición con el nombre 'Authorization Bearer'.<br>Para obtener la clave de autenticación,\r\n    *       ingresar al penel, sección de Configuración -> Sistema y luego en API.",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/proposal/create": {
            "post": {
                "tags": [
                    "Propuestas"
                ],
                "description": "Crear propuesta",
                "operationId": "30ef0710a1d9aa69a6d0a6489681c059",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddProposal"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TenantProposal"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/proposal/accept/{pid}/{did}": {
            "post": {
                "tags": [
                    "Propuestas"
                ],
                "description": "Aceptar propuesta",
                "operationId": "5d6c8ceb6844c395287f936686c58c90",
                "parameters": [
                    {
                        "name": "pid",
                        "in": "path",
                        "description": "Id de la propuesta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "did",
                        "in": "path",
                        "description": "Id del detalle de la propuesta",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "error": {
                                            "type": "boolean"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "AddProposal": {
                "title": "Schema de Alta de Propuestas",
                "properties": {
                    "products": {
                        "type": "array",
                        "items": {}
                    },
                    "quantities": {
                        "type": "array",
                        "items": {}
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "discount": {
                                    "type": "string"
                                },
                                "dues": {
                                    "type": "string"
                                },
                                "extra_id": {
                                    "type": "string"
                                },
                                "id": {
                                    "type": "string"
                                },
                                "init_amount": {
                                    "type": "string"
                                },
                                "installation": {
                                    "type": "string"
                                },
                                "installation_cost": {
                                    "type": "string"
                                },
                                "iva": {
                                    "type": "string"
                                },
                                "last_amount": {
                                    "type": "string"
                                },
                                "maintenance": {
                                    "type": "string"
                                },
                                "notes": {
                                    "type": "string"
                                },
                                "price": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "client_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "TenantProposal": {
                "title": "Schema de Propuesta",
                "properties": {
                    "id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}